2.ChatModels/requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
openai
langchain-openai

# Anthropic Integration
langchain-anthropic

# Google Gemini Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# FastAPI + Server
fastapi
uvicorn

# Environment Variables
python-dotenv

# ML utilities
numpy
scikit-learn

# HTTP utilities
requests
